tunnel further debugged testing services individually added
git-svn-id: https://jmg.devguard.com/svn/deops@104 a8feb433-1630-0410-ae02-e0276e0da4ce

+++ Services/Assist/VersionedCache.cs
@@ -187,7 +187,7 @@ namespace RiseOp.Services.Assist
 
                         FileRemoved.Invoke(vfile);
 
-                        if (vfile.Header != null)
+                        if (vfile.Header != null && vfile.Header.FileHash != null) // local sync doesnt use files
                             try { File.Delete(GetFilePath(vfile.Header)); }
                             catch { }
